Requirement update

+++ app/src/main/java/com/htnguyen/healthy/model/Items.java
@@ -3,9 +3,9 @@ package com.htnguyen.healthy.model;
 public class Items {
     private String date;
     private String description;
-    private int value;
+    private float value;
 
-    public Items(String date, String description, int value) {
+    public Items(String date, String description, float value) {
         this.date = date;
         this.description = description;
         this.value = value;
@@ -31,11 +31,11 @@ public class Items {
         this.description = description;
     }
 
-    public int getValue() {
+    public float getValue() {
         return value;
     }
 
-    public void setValue(int value) {
+    public void setValue(float value) {
         this.value = value;
     }
 }
